Refractor config & dbu locations
