+++ src/main.rs
@@ -1,10 +1,10 @@
 use actix_web::web;
 use once_cell::sync::Lazy;
 use sled::Db;
+use utils::config::Config;
 
-mod cfg;
-mod dbu;
 mod routes;
+mod utils;
 
 pub mod built_info {
     // The file has been placed there by the build script.
@@ -17,7 +17,8 @@ async fn p404() -> &'static str { "this resource does not exist." }
 
 #[actix_rt::main]
 async fn main() -> std::io::Result<()> {
-    let config = cfg::Config::load().await.unwrap();
+    let config = Config::load().await.unwrap();
+
     let port = config.port.clone();
     if !std::path::Path::new("./uploads").exists() {
         std::fs::create_dir_all("./uploads")?;
